size_t修正

+++ src/test/dummy_client.cc
@@ -40,10 +40,10 @@ DummyClient::DummyClient()
                               dummy_logger->debug("connection closed");
                               break;
                           } else if (meta && meta->offset > buf_s.size()) {
-                              buf_s.append(meta->offset - buf_s.size(), '\0');
+                              buf_s.append(static_cast<std::size_t>(meta->offset - buf_s.size()), '\0');
                               buf_s.append(buffer, rlen);
                           } else if (meta && meta->offset < buf_s.size()) {
-                              buf_s.replace(meta->offset, rlen, buffer, rlen);
+                              buf_s.replace(static_cast<std::size_t>(meta->offset), rlen, buffer, rlen);
                           } else {
                               buf_s.append(buffer, rlen);
                           }
